Add LoRaWAN implementation separately
I've moved the LoRaWAN implementation to a new class according to the
class diagram. It is now possible to change the payload while the task
is running and the new data will be sent every 5 minutes.

I kept the semaphore and task creation in the main.c file, but it should
probably be moved to a application.c file, that handles our freeRTOS
setup.